Show ahead/behind info for upstream and watch.
Follows (as near as I can tell) the standard git option of one remote and one or more merge
branches.

Additionally, adds a config value 'watch' for a branch, which can either be a space
separated list of watches, or multiple entries (i.e. multiple 'watch=' lines in the
config). Each watch entry is the remote and branch, separated by a colon.

For example, the default upstream, if converted to a watch, would be

  watch = origin:refs/heads/master

Note that this resembles a fetch spec, given the colon, but the meaning of both
sides is different. I don't know if this is an issue.

Further, the syntax gives the impression that we could watch arbitrary refs
of a remote, but that's not true either. Esp. since we strip the refs/heads off,
we are making an assumption about how the remote was fetched (namely the git
defaults).

A better syntax would probably just be a list of ref names, and let the user
do whatever they feel like. For that matter, there is no reason to not accept
arbitrary refspec, since that's what we pass to rev-list anyway, even though
having a commit-ish as a watch seems a bit silly.